correction for 3-hash_table_set.c

+++ 0x1A-hash_tables/2-key_index.c
@@ -13,7 +13,7 @@ unsigned long int key_index(const unsigned char *key, unsigned long int size)
 	unsigned long int hash;
 
 	hash = hash_djb2(key);
-	index = (unsigned long int) (hash % size);
+	index = hash % size;
 
 	return (index);
 }
